Clean up action history
I was using actions to dispatch changes for domains. Unfortunately,
that clogs up history a bit. I had a really hard time reasoning about
all of the edge cases, so I've added some internal, dev-only,
assertions to make sure things are never null.
