+++ package.json
@@ -27,7 +27,11 @@
   ],
   "license": "MIT",
   "jest": {
-    "coverageReporters": ["lcov", "text", "html"],
+    "coverageReporters": [
+      "lcov",
+      "text",
+      "html"
+    ],
     "setupTestFrameworkScriptFile": "./test/helpers/setup.js",
     "roots": [
       "<rootDir>/test",
@@ -45,6 +49,7 @@
     "babel-core": "^6.17.0",
     "babel-eslint": "^7.1.0",
     "babel-loader": "^6.2.9",
+    "babel-plugin-unassert": "^2.1.2",
     "babel-polyfill": "^6.22.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.16.0",
